+++ NEWS.md
@@ -1,3 +1,5 @@
+# epoxy (development version)
+
 # epoxy 0.1.1
 
 * `epoxy_transform_html()` now (again) returns a collapsed character string for