+++ NEWS.md
@@ -41,6 +41,10 @@
   of the first word in the expression. This is useful when you want to need to
   start a sentence with a variable that may contain more than one word. (#112)
 
+* The `.titlecase` inline transformer now coerces inputs to character with
+  `as.character()` before applying `tools::toTitleCase()`, since `toTitleCase()`
+  will throw an error for non-character inputs. (#112)
+
 # epoxy 0.1.1
 
 * `epoxy_transform_html()` now (again) returns a collapsed character string for